docs(DynamicPage): extract hook into separate file
