+++ packages/main/src/webComponents/DynamicPage/DynamicPage.stories.tsx
@@ -9,8 +9,7 @@ import exitFSIcon from '@ui5/webcomponents-icons/dist/exit-full-screen.js';
 import fullscreenIcon from '@ui5/webcomponents-icons/dist/full-screen.js';
 import { ThemingParameters } from '@ui5/webcomponents-react-base';
 import { clsx } from 'clsx';
-import type { RefObject } from 'react';
-import { useEffect, useRef, useState } from 'react';
+import { useRef } from 'react';
 import {
   Bar,
   Breadcrumbs,
@@ -29,6 +28,7 @@ import {
 } from '../..';
 import { DynamicPageHeader } from '../DynamicPageHeader/index.js';
 import { DynamicPageTitle } from '../DynamicPageTitle/index.js';
+import { useGetHeaderHeight } from './helpers.js';
 import type { DynamicPageDomRef } from './index.js';
 import { DynamicPage } from './index.js';
 
@@ -142,40 +142,6 @@ export const Default: Story = {
       </DynamicPage>
     );
   }
-};
-
-export const useGetHeaderHeight = (dynamicPageRef: RefObject<DynamicPageDomRef & { shadowRoot: ShadowRoot }>) => {
-  const [headerHeight, setHeaderHeight] = useState<undefined | number>(undefined);
-  useEffect(() => {
-    const headerObserver = new ResizeObserver(([header]) => {
-      setHeaderHeight(header.contentRect.height);
-    });
-    if (dynamicPageRef.current) {
-      // wait for the custom element to be defined (adjust the tag-name if you're using the scoping feature)
-      void customElements.whenDefined('ui5-dynamic-page').then(() => {
-        const { shadowRoot } = dynamicPageRef.current;
-
-        // wait for the shadowRoot to be populated
-        const shadowRootObserver = new MutationObserver(() => {
-          const header = shadowRoot.querySelector('header');
-          if (header) {
-            shadowRootObserver.disconnect();
-            headerObserver.observe(header);
-          }
-        });
-
-        if (shadowRoot && shadowRoot.childElementCount) {
-          headerObserver.observe(shadowRoot.querySelector('header'));
-        } else {
-          shadowRootObserver.observe(shadowRoot, { childList: true });
-        }
-      });
-    }
-    return () => {
-      headerObserver.disconnect();
-    };
-  }, []);
-  return headerHeight;
 };
 
 export const StickyContentHeaders: Story = {
