+++ packages/main/src/webComponents/DynamicPage/helpers.ts
@@ -1,0 +1,37 @@
+import type { RefObject } from 'react';
+import { useEffect, useState } from 'react';
+import type { DynamicPageDomRef } from './index.js';
+
+export const useGetHeaderHeight = (dynamicPageRef: RefObject<DynamicPageDomRef & { shadowRoot: ShadowRoot }>) => {
+  const [headerHeight, setHeaderHeight] = useState<undefined | number>(undefined);
+  useEffect(() => {
+    const headerObserver = new ResizeObserver(([header]) => {
+      setHeaderHeight(header.contentRect.height);
+    });
+    if (dynamicPageRef.current) {
+      // wait for the custom element to be defined (adjust the tag-name if you're using the scoping feature)
+      void customElements.whenDefined('ui5-dynamic-page').then(() => {
+        const { shadowRoot } = dynamicPageRef.current;
+
+        // wait for the shadowRoot to be populated
+        const shadowRootObserver = new MutationObserver(() => {
+          const header = shadowRoot.querySelector('header');
+          if (header) {
+            shadowRootObserver.disconnect();
+            headerObserver.observe(header);
+          }
+        });
+
+        if (shadowRoot && shadowRoot.childElementCount) {
+          headerObserver.observe(shadowRoot.querySelector('header'));
+        } else {
+          shadowRootObserver.observe(shadowRoot, { childList: true });
+        }
+      });
+    }
+    return () => {
+      headerObserver.disconnect();
+    };
+  }, []);
+  return headerHeight;
+};
